requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
langgraph==0.2.14
faiss-cpu==1.8.0
sentence-transformers==3.0.1
groq==0.11.0
httpx==0.27.2
python-dotenv==1.0.1